refactor(cli): extract credential source to lib/credential.ts; keep static auth path unchanged
Follow-up cleanup on the credential-command auth change so it fits the existing
direct-mode auth instead of bolting a parallel path onto it.

- Static env key no longer routes through the credential-source machinery. A
  `direct` connection with no `auth` block passes its env key as `apiKey`
  exactly as before (the AI SDK sends the wire's conventional header); the
  injecting fetch + placeholder apiKey engage only when an `auth` block is
  configured. Removed `staticCredentialSource`, which existed only to force the
  static path through the new seam and re-set an identical header.
- Moved the CredentialSource machinery (command minting, ExecCredential/raw
  parsing, expiry caching, refresh) out of lib/env.ts into lib/credential.ts;
  it is not process.env work. The `env` kind reads via env.ts's new
  `readEnvCredentialVar` seam, so env.ts stays the sole `process.env` reader.
  env.ts's addition drops from ~215 to ~52 lines.
- Trimmed the excessive comments across the change to a reasonable density,
  keeping the load-bearing rationale (placeholder apiKey, managed-helper never
  auto-run, raw vs ExecCredential).

Credential unit tests moved to lib/credential.test.ts (added to the ESLint
process.env exemption, with rationale). OpenSpec proposal/tasks/design updated
to match the new file layout and the static-key path.

typecheck / lint / format clean; affected suites 128 pass / 0 fail.
